requirements.txt
numpy
torch
torchvision
pillow
PyYaml
opencv-python
matplotlib
tensorboard